1) + inserting animation; 2) + forwarding scroll-view delegate method; 3) ~ couple fixes.

+++ MMGridView/MMGridLayout.m
@@ -231,10 +231,11 @@
     CGSize pageSize = scrollView.frame.size;
     switch (layout) {
         case MMGridLayoutPagedHorizontal: {
-            return (NSUInteger) (floor((scrollView.contentOffset.x - pageSize.width / 2) / pageSize.width) + 1);
+            //  cuz (NSUInteger)(double)(-1) == 0
+            return (NSUInteger) (NSInteger) (floor((scrollView.contentOffset.x - pageSize.width / 2) / pageSize.width) + 1.);
         }
         case MMGridLayoutPagedVertical : {
-            return (NSUInteger) (floor((scrollView.contentOffset.y - pageSize.height / 2) / pageSize.height) + 1);
+            return (NSUInteger) (NSInteger) (floor((scrollView.contentOffset.y - pageSize.height / 2) / pageSize.height) + 1.);
         }
         default: {
             return 0;
